Move the home folder to pages
