+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import Main from '@/components/home/Main';
+import Main from '@/pages/home/Main';
 
 export default function Home() {
     return <Main />;
